only delete exercises if they match the custom bitmap value

+++ src/main/java/com/ooush/api/service/exercise/ExerciseServiceImpl.java
@@ -86,7 +86,9 @@ public class ExerciseServiceImpl implements ExerciseService {
 		UserExercise userExercise = userExerciseRepository.findByUserAndExerciseDayAndExercise(currentLoggedInUser, exerciseDay, exercise);
 		if (userExercise != null) {
 			userExerciseRepository.delete(userExercise);
-			exerciseRepository.delete(exercise);
+			if (exercise.isCustomExercise()) {
+				exerciseRepository.delete(exercise);
+			}
 		}
 	}
